introduce kubecli/websocket.go
Things that are related to websocket handling are moved from
kubecli/vmi.go to kubecli/websocket.go.
